transfer methods to new classes
